government menu

+++ project-group-50-phase1-2/Phase 1/Stronghold/src/main/java/model/ObjectsPackage/People/Person.java
@@ -10,7 +10,7 @@ public abstract class Person extends Objects {
     private int life;
     private int speed;
     private int starving;
-    private int income;
+    private double income;
 
     protected Person(boolean isSoldier, int life, int speed, User owner) {
         super(ObjectType.PERSON, owner);
@@ -60,11 +60,11 @@ public abstract class Person extends Objects {
         this.starving = starving;
     }
 
-    public int getIncome() {
+    public double getIncome() {
         return income;
     }
 
-    public void setIncome(int income) {
+    public void setIncome(double income) {
         this.income = income;
     }
